requirements.txt
Django==4.0.4
django-model-utils==4.2.0
djangorestframework==3.13.1